refactor(docs): extract catalog/settings/paths + coverage into lib/reference-pages.ts
Final docs-generate split PR. Pure relocation of the remaining reference
renderers + the coverage report out of scripts/docs/generate.ts:
buildCoverageReport, renderPluginCatalog (core-plugins.md), renderSettingsReference
(settings.md), renderRuntimePathsReference (runtime-paths.md), and the three
coverage-only consts (llmBundleFiles/docsSnippetFiles/publicSlotNames, now
module-internal). Function bodies byte-for-byte unchanged; generate.ts imports
the four renderers back. Dropped the now-unused source-scan, api-docs,
escapeHtml/generatedPageNote, and three source-extraction imports.

generate.ts is now a 310-line invocation-only entry: imports + path consts +
versionLine + the apiReferenceGroups orchestration + the writeStableFile/bundle
calls (down from 2,585 at the start of the docs-generate work).

Verified: docs:generate reference output byte-identical AND the full docs diff
(coverage.json + every generated page + marker blocks) identical to the parent
branch. eslint clean.
